requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
tldextract==5.1.1 